Move whitelist middleware to its own module
